pyproject.toml
[project]
name = "livekit-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "livekit>=1.0.16",
    "livekit-agents[inworld]>=1.2.14",
    "livekit-plugins-cartesia>=1.2.14",
    "livekit-plugins-deepgram>=1.2.14",
    "livekit-plugins-google>=1.2.14",
    "livekit-plugins-noise-cancellation>=0.2.5",
    "livekit-plugins-silero>=1.2.14",
    "livekit-plugins-turn-detector>=1.2.14",
    "python-dotenv>=1.1.1",
]
